Add Lock to lock a resource
This is based on redlock and https://redis.io/topics/distlock
I did not want to include that gem that is based on a multi instances
redis.
The documentation explitely says that if we are ok having maybe 2 locks
during a failover this approach is enough.

We might double deploy to an app in this scenario
1. A tries to deploy app X and creates a lock.
2. The lock was not transfered from master to follower
3. Redis failover and the follower is becoming a master
4. B tries to deploy app X and the locks is created on the new master.