Simplify signup complete process
`SignupComplete#notify_user_of_success` always returns early because
`user` is always `nil`.  See here:
https://github.com/heroku/slash-heroku/blob/master/app/models/signup_complete.rb#L15

The `user` is always `nil` because we try to find it by `slack_user_id`,
but we pass it the `User#id` (the unique `id` from the database).
See here:
https://github.com/heroku/slash-heroku/blob/master/app/controllers/sessions_controller.rb#L64
https://github.com/heroku/slash-heroku/blob/master/app/jobs/signup_complete_job.rb#L15

So that means the `you're all set. :tada:` message was never displayed.
Also, we now want to display both the Heroku and GitHub login statuses
to show the user his current state. That's why the `SignupComplete`
class was removed.

The only thing `SignupCompleteJob` does now is create a
`CommandExecutorJob`, see here:
https://github.com/heroku/slash-heroku/blob/master/app/jobs/signup_complete_job.rb#L16

So it makes sense to delete the `SignupCompleteJob` and run a
`CommandExecutorJob` directly.