Only check if heroku is configured in 1 place
We check `current_user.heroku_token` before a command is created and
`heroku_configured?` before a command is run.

We can check `heroku_token` (which checks for the token validity and
refreshes it if needed) in the `heroku_configured?` method, so we only
do that in 1 place.

Checking `current_user.heroku_token` before creating a command also has
the side effect of creating a command without a user if the token has
expired. It means the `login` command will tell the user he has to sign
in to Heroku *and* GitHub, even if the user has a valid GitHub token
already.